release: v1.1.0 — developer experience, showcase portfolio, dataset hub
- Complete README overhaul: hero → insights → quick start → datasets → features → scouting → gallery → architecture
- docs/getting_started.md: 5-minute tutorial, first useful insight under 5 minutes
- docs/gallery.md: 10 showcase studies with charts, SQL, key findings
- docs/onboarding_audit.md: 4-persona cold-start audit, 2 critical / 9 major / 6 minor issues
- CHANGELOG.md v1.1.0 entry with full feature list and known issues
- FEATURES.md: complete feature store reference
- Dataset hub: midwicket.datasets.load_dataset() for 12 Cricsheet competitions
- Feature store: 10 builders (pressure index, BQR, intent score, VBR, MCS, expected runs/wickets, batting form, death metrics, venue-adjusted form)
- Scouting reports: midwicket.scouting_report() with name alias resolution
- 25 showcase scripts in examples/; 10 executed with full PNG charts and walkthroughs in docs/showcases/
- Independent verification report: 1,239 matches, 100% success rate, 294,757 deliveries
- Schema validation report: int16/int32 overflow fixes confirmed, retirement fix confirmed, temporal leakage confirmed clean
- 34 new tests for reliability and v2 features/datasets

+++ midwicket/schema/v1.py
@@ -77,15 +77,15 @@ BALL_EVENT_SCHEMA = pa.schema([
     
     # --- State (When) ---
     ('inning', pa.int8()),
-    ('over', pa.int8()),
+    ('over', pa.int16()),
     ('ball', pa.int8()),
     
     # --- Actors (IDs from Registry) ---
     ('batter_id', pa.int32()),
     ('bowler_id', pa.int32()),
     ('non_striker_id', pa.int32()),
-    ('batting_team_id', pa.int16()),
-    ('bowling_team_id', pa.int16()),
+    ('batting_team_id', pa.int32()),
+    ('bowling_team_id', pa.int32()),
     
     # --- Metrics (What Happened) ---
     # Upcasted to int32 to prevent DuckDB SUM() overflow
